ACT4 - NiverMtz -  User.tsx applying componentization
